+++ Account Classes/AccountCurrent.cs
@@ -11,7 +11,7 @@ namespace GoodBankNS.AccountClasses
 	public class AccountCurrent : Account
 	{
 		public override AccountType AccType { get => AccountType.Current; }
-		public override int Balance { get; set; }
+		public override double Balance { get; set; }
 
 		/// <summary>
 		/// Создание счета на основе введенных данных
